Rename old molecules module
